HBX-3080: Refactor the Gradle integration tests to factor out common code   - Move test method 'JpaDefaultTest#testJpaDefault()' to test class 'GenerateJavaTest' and delete 'JpaDefaultTest'   - Move test method 'NoAnnotationsTest#testNoAnnotations()' to test class 'GenerateJavaTest' and delete 'NoAnnotationsTest'   - Move test method 'NoGenerics#testNoGenerics()' to test class 'GenerateJavaTest' and delete 'NoGenerics'   - Move test method 'UseGenerics#testUseGenerics()' to test class 'GenerateJavaTest' and delete 'UseGenerics'   - Create instance variable 'TestTemplate#gradleCommandToExecute' with getter/setter   - Create method 'TestTemplate#createProjectAndExecuteGradleCommand()' that uses the stored command   - Use method 'createProjectAndExecuteGradleCommand()' in the test methods of 'GenerateJavaTest'   - Move class 'TutorialTest' to package 'org.hibernate.tool.gradle'
