+++ gradle/plugin/src/functionalTest/java/org/hibernate/tool/gradle/GenerateJavaTest.java
@@ -17,32 +17,133 @@
  */
 package org.hibernate.tool.gradle;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
-import java.io.IOException;
 import java.nio.file.Files;
 
-import org.gradle.testkit.runner.BuildResult;
-import org.hibernate.tool.gradle.test.func.utils.FuncTestConstants;
-import org.hibernate.tool.gradle.test.func.utils.FuncTestTemplate;
 import org.hibernate.tool.it.gradle.TestTemplate;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 class GenerateJavaTest extends TestTemplate {
 
-	@Test
-	public void testPackageName() throws Exception {
+	@BeforeEach
+	public void beforeEach() {
+		setGradleCommandToExecute("generateJava");
 		setDatabaseCreationScript(new String[] {
 				"create table PERSON (ID int not null, NAME varchar(20), primary key (ID))"
 		});
+	}
+
+	@Test
+	public void testJpaDefault() throws Exception {
+		createProjectAndExecuteGradleCommand();
+		File generatedOutputFolder = new File(getProjectDir(), "app/generated-sources");
+		assertTrue(generatedOutputFolder.exists());
+		assertTrue(generatedOutputFolder.isDirectory());
+		assertEquals(1, generatedOutputFolder.list().length);
+		File generatedPersonJavaFile = new File(generatedOutputFolder, "Person.java");
+		assertTrue(generatedPersonJavaFile.exists());
+		assertTrue(generatedPersonJavaFile.isFile());
+		String generatedPersonJavaFileContents = new String(
+				Files.readAllBytes(generatedPersonJavaFile.toPath()));
+		assertTrue(generatedPersonJavaFileContents.contains("import jakarta.persistence.Entity;"));
+		assertTrue(generatedPersonJavaFileContents.contains("public class Person "));
+	}
+
+	@Test
+	public void testNoAnnotations() throws Exception {
+		setHibernateToolsExtensionSection(
+				"hibernateTools { \n" +
+						"  generateAnnotations=false \n" +
+						"}"
+		);
+		createProjectAndExecuteGradleCommand();
+		File generatedOutputFolder = new File(getProjectDir(), "app/generated-sources");
+		assertTrue(generatedOutputFolder.exists());
+		assertTrue(generatedOutputFolder.isDirectory());
+		assertEquals(1, generatedOutputFolder.list().length);
+		File generatedPersonJavaFile = new File(generatedOutputFolder, "Person.java");
+		assertTrue(generatedPersonJavaFile.exists());
+		assertTrue(generatedPersonJavaFile.isFile());
+		String generatedPersonJavaFileContents = new String(
+				Files.readAllBytes(generatedPersonJavaFile.toPath()));
+		assertFalse(generatedPersonJavaFileContents.contains("import jakarta.persistence.Entity;"));
+		assertTrue(generatedPersonJavaFileContents.contains("public class Person "));
+	}
+
+	@Test
+	public void testNoGenerics() throws Exception {
+		setDatabaseCreationScript(new String[] {
+				"create table PERSON (ID int not null,  NAME varchar(20), primary key (ID))",
+				"create table ITEM (ID int not null,  NAME varchar(20), OWNER_ID int not null, " +
+						"   primary key (ID), foreign key (OWNER_ID) references PERSON(ID))"
+		});
+		setHibernateToolsExtensionSection(
+				"hibernateTools { \n" +
+						"  useGenerics=false \n" +
+						"}"
+		);
+		createProjectAndExecuteGradleCommand();
+		executeGradleCommand("generateJava");
+		File generatedOutputFolder = new File(getProjectDir(), "app/generated-sources");
+		assertTrue(generatedOutputFolder.exists());
+		assertTrue(generatedOutputFolder.isDirectory());
+		assertEquals(2, generatedOutputFolder.list().length);
+		File generatedPersonJavaFile = new File(generatedOutputFolder, "Person.java");
+		assertTrue(generatedPersonJavaFile.exists());
+		assertTrue(generatedPersonJavaFile.isFile());
+		String generatedPersonJavaFileContents = new String(
+				Files.readAllBytes(generatedPersonJavaFile.toPath()));
+		assertTrue(generatedPersonJavaFileContents.contains("public class Person "));
+		assertFalse(generatedPersonJavaFileContents.contains("Set<Item>"));
+		File generatedItemJavaFile = new File(generatedOutputFolder, "Item.java");
+		assertTrue(generatedItemJavaFile.exists());
+		assertTrue(generatedItemJavaFile.isFile());
+		String generatedItemJavaFileContents = new String(
+				Files.readAllBytes(generatedItemJavaFile.toPath()));
+		assertTrue(generatedItemJavaFileContents.contains("public class Item "));
+	}
+
+	@Test
+	public void testUseGenerics() throws Exception {
+		setDatabaseCreationScript(new String[] {
+				"create table PERSON (ID int not null,  NAME varchar(20), primary key (ID))",
+				"create table ITEM (ID int not null,  NAME varchar(20), OWNER_ID int not null, " +
+						"   primary key (ID), foreign key (OWNER_ID) references PERSON(ID))"
+		});
+		createProjectAndExecuteGradleCommand();
+		executeGradleCommand("generateJava");
+		File generatedOutputFolder = new File(getProjectDir(), "app/generated-sources");
+		assertTrue(generatedOutputFolder.exists());
+		assertTrue(generatedOutputFolder.isDirectory());
+		assertEquals(2, generatedOutputFolder.list().length);
+		File generatedPersonJavaFile = new File(generatedOutputFolder, "Person.java");
+		assertTrue(generatedPersonJavaFile.exists());
+		assertTrue(generatedPersonJavaFile.isFile());
+		String generatedPersonJavaFileContents = new String(
+				Files.readAllBytes(generatedPersonJavaFile.toPath()));
+		assertTrue(generatedPersonJavaFileContents.contains("public class Person "));
+		assertTrue(generatedPersonJavaFileContents.contains("Set<Item>"));
+		File generatedItemJavaFile = new File(generatedOutputFolder, "Item.java");
+		assertTrue(generatedItemJavaFile.exists());
+		assertTrue(generatedItemJavaFile.isFile());
+		String generatedItemJavaFileContents = new String(
+				Files.readAllBytes(generatedItemJavaFile.toPath()));
+		assertTrue(generatedItemJavaFileContents.contains("public class Item "));
+	}
+
+	@Test
+	public void testPackageName() throws Exception {
 		setHibernateToolsExtensionSection(
 				"hibernateTools { \n" +
 				"  packageName = 'foo.model' \n" +
 				"}"
 		);
-		createProject();
-		executeGradleCommand("generateJava");
+		createProjectAndExecuteGradleCommand();
 		File generatedSourcesFolder = new File(getProjectDir(), "app/generated-sources");
 		assertTrue(generatedSourcesFolder.exists());
 		assertTrue(generatedSourcesFolder.isDirectory());
@@ -50,6 +151,5 @@ class GenerateJavaTest extends TestTemplate {
 		assertTrue(fooFile.exists());
 		assertTrue(fooFile.isFile());
 	}
-
 
 }
