+++ gradle/plugin/src/functionalTest/java/org/hibernate/tool/gradle/TutorialTest.java
@@ -1,16 +1,10 @@
-package org.hibernate.tool.gradle.java;
+package org.hibernate.tool.gradle;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.File;
-import java.nio.file.Files;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.ResultSet;
-import java.sql.Statement;
-import java.util.List;
 
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
@@ -18,10 +12,10 @@ import org.junit.jupiter.api.Test;
 
 import org.hibernate.tool.it.gradle.TestTemplate;
 
-public class JpaDefaultTest extends TestTemplate {
+public class TutorialTest extends TestTemplate {
 	
 	@Test
-	public void testJpaDefault() throws Exception {
+	public void testTutorial() throws Exception {
 		setDatabaseCreationScript(new String[] {
 				"create table PERSON (ID int not null, NAME varchar(20), primary key (ID))",
 				"insert into PERSON values (1, 'foo')"
@@ -35,10 +29,6 @@ public class JpaDefaultTest extends TestTemplate {
 		File generatedPersonJavaFile = new File(generatedOutputFolder, "Person.java");
 		assertTrue(generatedPersonJavaFile.exists());
 		assertTrue(generatedPersonJavaFile.isFile());
-		String generatedPersonJavaFileContents = new String(
-				Files.readAllBytes(generatedPersonJavaFile.toPath()));
-		assertTrue(generatedPersonJavaFileContents.contains("import jakarta.persistence.Entity;"));
-		assertTrue(generatedPersonJavaFileContents.contains("public class Person "));
 	}
 	
 }
